moved filemetadata into it's own model file
