+++ core/src/model/file_metadata.rs
@@ -1,8 +1,9 @@
-use crate::model::crypto::{FolderAccessInfo, UserAccessInfo};
+use crate::model::crypto::{FolderAccessInfo, UserAccessInfo, SignedValue, EncryptedValueWithNonce};
 use serde::{Deserialize, Serialize};
 use std::clone::Clone;
 use std::collections::HashMap;
 use uuid::Uuid;
+use crate::model::account::Username;
 
 #[derive(Clone, PartialEq, Debug, Deserialize, Serialize, Copy)]
 pub enum FileType {
@@ -47,4 +48,19 @@ pub struct ClientFileMetadata {
 
     /// True if the user attempted to delete this file locally. Once the server also deletes this file, the content and the associated metadata are deleted locally.
     pub deleted: bool,
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+pub struct FileMetadata {
+    pub id: Uuid,
+    pub file_type: FileType,
+    pub parent: Uuid,
+    pub name: String,
+    pub owner: String,
+    pub signature: SignedValue,
+    pub metadata_version: u64,
+    pub content_version: u64,
+    pub deleted: bool,
+    pub user_access_keys: HashMap<Username, UserAccessInfo>,
+    pub folder_access_keys: EncryptedValueWithNonce,
 }
