+++ core/src/model/mod.rs
@@ -1,6 +1,6 @@
 pub mod account;
 pub mod api;
-pub mod client_file_metadata;
+pub mod file_metadata;
 pub mod crypto;
 pub mod state;
 pub mod work_unit;
