+++ server/src/file_service.rs
@@ -2,7 +2,7 @@ use crate::files_db;
 use crate::index_db;
 use crate::ServerState;
 use lockbook_core::model::api::*;
-use lockbook_core::model::client_file_metadata::FileType;
+use lockbook_core::model::file_metadata::FileType;
 
 pub fn username_is_valid(username: &str) -> bool {
     username
